requirements.txt
asgiref==3.3.1
Django==3.1.7
pkg-resources==0.0.0
pytz==2021.1
sqlparse==0.4.1

requests~=2.25.1
bs4~=0.0.1
beautifulsoup4~=4.9.3
djangorestframework~=3.12.2